Reorganise migrate command into a its own struct & add --force flag
