+++ cmd/ctl/pkg/upgrade/migrate/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["migrate.go"],
+    srcs = [
+        "command.go",
+        "migrator.go",
+    ],
     importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/upgrade/migrate",
     visibility = ["//visibility:public"],
     deps = [
